requirements.txt
adal==1.2.7
applicationinsights==0.11.10
argcomplete==1.12.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
async-generator==1.10
attrs==21.4.0
azure-common==1.1.28
azure-core==1.21.1
azure-graphrbac==0.61.1
azure-mgmt-authorization==0.61.0
azure-mgmt-containerregistry==8.2.0
azure-mgmt-core==1.3.0
azure-mgmt-keyvault==9.3.0
azure-mgmt-resource==20.1.0
azure-mgmt-storage==19.1.0
azure-storage-blob==12.11.0
azureml-core==1.40.0
azureml-telemetry==1.40.0
azureml-widgets==1.40.0
backcall==0.2.0
backports.tempfile==1.0
backports.weakref==1.0.post1
bcrypt==3.2.0
bleach==4.1.0
certifi==2020.6.20
cffi==1.15.0
charset-normalizer==2.0.12
colorama==0.4.4
contextlib2==21.6.0
cryptography==36.0.2
cycler==0.11.0
dataclasses==0.8
decorator==5.1.1
defusedxml==0.7.1
docker==5.0.3
entrypoints==0.4
fairness-indicators==0.1.2
humanfriendly==10.0
idna==3.3
importlib-metadata==4.8.3
ipykernel==5.5.6
ipython==7.16.3
ipython-genutils==0.2.0
ipywidgets==7.7.0
isodate==0.6.1
jedi==0.17.2
jeepney==0.7.1
Jinja2==3.0.3
jmespath==0.10.0
jsonpickle==2.1.0
jsonschema==3.2.0
jupyter-client==7.1.2
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.1.0
kiwisolver==1.3.1
knack==0.9.0
MarkupSafe==2.0.1
matplotlib==3.3.4
mistune==0.8.4
msal==1.17.0
msal-extensions==0.3.1
msrest==0.6.21
msrestazure==0.6.4
nbclient==0.5.9
nbconvert==6.0.7
nbformat==5.1.3
ndg-httpsclient==0.5.1
nest-asyncio==1.5.5
notebook==6.4.10
numpy==1.19.5
oauthlib==3.2.0
packaging==21.3
pandas==1.1.5
pandocfilters==1.5.0
paramiko==2.10.3
parso==0.7.1
pathspec==0.9.0
pickleshare==0.7.5
Pillow==8.4.0
pkginfo==1.8.2
portalocker==2.4.0
prometheus-client==0.14.1
prompt-toolkit==3.0.29
pyasn1==0.4.8
pycparser==2.21
Pygments==2.11.2
PyJWT==2.3.0
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==3.0.7
pyreadline==2.1
pyrsistent==0.18.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.1
pywin32==227
pywinpty==1.1.6
PyYAML==6.0
pyzmq==22.3.0
requests==2.27.1
requests-oauthlib==1.3.1
SecretStorage==3.3.1
Send2Trash==1.8.0
six==1.16.0
tabulate==0.8.9
tensorflow-metadata==0.30.0
terminado==0.12.1
testpath==0.6.0
tornado==6.1
traitlets==4.3.3
typing_extensions==4.1.1
urllib3==1.26.7
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.1
widgetsnbextension==3.6.0
wincertstore==0.2
zipp==3.6.0
